Relocate memory specific data to own file
allows easier handling for memory interactions
