+++ PKHeX.Core/Editing/CommonEdits.cs
@@ -579,7 +579,7 @@ namespace PKHeX.Core
         {
             pk.OT_Memory = 2;
             pk.OT_Affection = 0;
-            pk.OT_Feeling = Legal.GetRandomFeeling(pk.OT_Memory);
+            pk.OT_Feeling = Memories.GetRandomFeeling(pk.OT_Memory);
             pk.OT_Intensity = 1;
             pk.OT_TextVar = pk.XY ? 43 : 27; // riverside road : battling spot
         }
@@ -589,7 +589,7 @@ namespace PKHeX.Core
             // for lack of better randomization :)
             pk.OT_Memory = 63;
             pk.OT_Intensity = 6;
-            pk.OT_Feeling = Legal.GetRandomFeeling(pk.OT_Memory);
+            pk.OT_Feeling = Memories.GetRandomFeeling(pk.OT_Memory);
         }
     }
 }
