+++ PKHeX.Core/Game/GameStrings/GameInfo.cs
