+++ PKHeX.Core/Game/GameStrings/MemoryArgType.cs
@@ -1,0 +1,12 @@
+﻿namespace PKHeX.Core
+{
+    public enum MemoryArgType
+    {
+        None,
+        GeneralLocation,
+        SpecificLocation,
+        Species,
+        Move,
+        Item,
+    }
+}
