+++ PKHeX.Core/Legality/Core.cs
@@ -2239,16 +2239,5 @@ namespace PKHeX.Core
                     return LanguageID.English;
             }
         }
-
-        public static int GetRandomFeeling(int memory, int max = 24)
-        {
-            var bits = MemoryFeelings[memory];
-            while (true)
-            {
-                int feel = Util.Rand.Next(0, max);
-                if ((bits & (1 << feel)) != 0)
-                    return feel;
-            }
-        }
     }
 }
