+++ PKHeX.Core/MysteryGifts/WC6.cs
@@ -363,14 +363,14 @@ namespace PKHeX.Core
                 pk.OT_Memory = 3;
                 pk.OT_TextVar = 9;
                 pk.OT_Intensity = 1;
-                pk.OT_Feeling = Legal.GetRandomFeeling(pk.OT_Memory, 10); // 0-9
+                pk.OT_Feeling = Memories.GetRandomFeeling(pk.OT_Memory, 10); // 0-9
             }
             else
             {
                 pk.HT_Memory = 3;
                 pk.HT_TextVar = 9;
                 pk.HT_Intensity = 1;
-                pk.HT_Feeling = Legal.GetRandomFeeling(pk.HT_Memory, 10); // 0-9
+                pk.HT_Feeling = Memories.GetRandomFeeling(pk.HT_Memory, 10); // 0-9
                 pk.HT_Friendship = pk.OT_Friendship;
             }
             pk.IsNicknamed = IsNicknamed;
